refactor protobufs
